T6-99 put toolbar test
