+++ src/components/textEditor/SlateEditor.tsx
@@ -2,8 +2,14 @@
 
 import React from 'react';
 
+import Toolbar from './Toolbar';
+
 function SlateEditor() {
-  return <div>Slate Editor</div>;
+  return (
+    <div>
+      <Toolbar />
+    </div>
+  );
 }
 
 export default SlateEditor;
